+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Change Log
 ==========
 
-## Version 2.0.0 (2021-03-12) 🚀
+## Version 2.0.0 *(2021-05-17)* 🚀
 
 * 🆕 Previous boring custom layout system removed. And view-binding supported custom layout system added. See `CarouselListener` in the sample app for details. 🎉
 * 🆕 The carousel is now supported Infinite ∞ looping 🥳! It's default now. You can disable it by setting `infiniteCarousel` to `false`.
@@ -12,7 +12,9 @@ Change Log
 * 🆕 You can now use `setData()` to set `CarouselItem` list. And `addData()` to add single or multiple items to the carousel. If you set/add data only once, then you can continue using `addData()` method.
 * 🆕 You can now get the carousel items using `getData()` method.
 * 🆕 Sample app re-designed. Various use cases and inspiring 💡 examples are added. 😎
-* ⚠️ `OnItemClickListener` renamed to `CarouselListener` and two new method (`onCreateViewHolder()` and `onBindViewHolder()`) added for the latest custom view feature.  `setOnItemClickListener()` also renamed to `setCarouselListener()`.
+* ⚠️ `OnItemClickListener` renamed to `CarouselListener` and two new method (`onCreateViewHolder()` and `onBindViewHolder()`) added for the latest custom view feature. `setOnItemClickListener()` also renamed to `setCarouselListener()`.
+* ⚠️ `CarouselListener` (previously `OnItemClickListener`) package changed from `org.imaginativeworld.whynotimagecarousel` to `org.imaginativeworld.whynotimagecarousel.listener`.
+* ⚠️ `CarouselItem` package changed from `org.imaginativeworld.whynotimagecarousel` to `org.imaginativeworld.whynotimagecarousel.model`.
 * 🛠️ New parameters `position` and `carouselItem` add to `onScrolled` method in  `CarouselOnScrollListener`.
 * 🛠️ The default carousel background changed to transparent. You can always change the background of the carousel using `carouselBackground`. The previous value was `#333333`.
 * 🛠️ Left and right navigation default button icon replaced with a rounded chevron icon.